move game settings into src
this was done because the script needs to be callable from robot-compile
which is not located in this package.
Therefore the best way is to move the scripts functionality into src
while keeping an executable here which simply calls the content from src
